% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateExampleData.R
\name{generate_example_data}
\alias{generate_example_data}
\title{generate an example four dimensional data tensor}
\usage{
generate_example_data(N = 200, L = 500, M = 16, paramT = 3, sparsity = 0.4)
}
\arguments{
\item{N}{integer number of samples, default is 200.}

\item{L}{integer number of genes, default is 500.}

\item{M}{integer number of time points, default is 16.}

\item{paramT}{integer number of tissues, default is 3.}

\item{sparsity, }{default is 0.4.}
}
\value{
list consisting of N by L by M by paramT dataTensor and the
four matrices A B D X that generate it.
}
\description{
generate an example four dimensional data tensor
}
