% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_matrix_or_tensor.R
\name{matrixToTensor}
\alias{matrixToTensor}
\title{matrixToTensor converts 2D tensor to 4D  array suitable for SDA4D.}
\usage{
matrixToTensor(input_matrix)
}
\arguments{
\item{input_matrix}{numeric matrix}
}
\value{
four dimensional tensor with entries from input_matrix.
}
\description{
matrixToTensor converts a matrix of dimension N by L to a tensor
 of dimension N by L by 1 by 1, as is required to run SDA4D.
}
\examples{
matrixToTensor(matrix(rnorm(50),25,2))
}
